eslint and streaks

+++ apps/browser-extension/custom-vite-plugins.ts
@@ -1,6 +1,8 @@
 import fs from 'fs';
 import { resolve } from 'path';
 import type { PluginOption } from 'vite';
+import type { OutputOptions, InputOptions } from 'rollup';
+
 
 // plugin to remove dev icons from prod build
 export function stripDevIcons(isDev: boolean) {
@@ -11,8 +13,9 @@ export function stripDevIcons(isDev: boolean) {
 		resolveId(source: string) {
 			return source === 'virtual-module' ? source : null;
 		},
-		renderStart(outputOptions: any, inputOptions: any) {
+		renderStart(outputOptions: OutputOptions, _inputOptions: InputOptions) {
 			const outDir = outputOptions.dir;
+			if (!outDir) return;
 			fs.rm(resolve(outDir, 'dev-icon-32.png'), () =>
 				console.log(`Deleted dev-icon-32.png from prod build`)
 			);
